Renamed ActivityResultHandler to ITMCoActivityResult
